applied Jon's suggestions
