+++ pyproject.toml
@@ -16,7 +16,6 @@ dynamic = ["version"]
 dependencies = [
     'pydantic<2.0',
     'jsonschema',
-    'packaging',
     'typing_extensions; python_version < "3.8"',
     'typing_extensions < 4.6.0; python_version < "3.10"' # we will need to unpin this when pydantic releases a new version
 ]
