+++ tests/resources/ephys_data_description/data_description_0.6.2.json
@@ -8,9 +8,8 @@
         }
     ],
     "investigators": [
-        "Galen Lynch",
-        "Yoni Browning",
-        "Josh Siegle"
+        "John Doe",
+        "Mary Smith"
     ],
     "modality": [
         {
